test: rename reservedconn reconnect packages to descriptive names
