+++ go/test/endtoend/vtgate/reservedconn/tabletchange/main_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package reservedconn
+package tabletchange
 
 import (
 	"flag"
